Fix the Scunthorpe problem.
The Scunthorpe problem is a problem where a filter blocks an innocent
message due to the presence of an illegal word included in a non-profane word. A common example is in the word "Scunthorpe" (a town in England), which contains the word "cunt", and, thus, most filters reject the word as profane. This problem is fixed by adding word boundaries (\b characters) before and after the
word.